feat: add class-is support for connections
