+++ package.json
@@ -54,6 +54,7 @@
   "dependencies": {
     "async": "^2.6.1",
     "big.js": "^5.1.2",
+    "class-is": "^1.1.0",
     "debug": "^3.1.0",
     "err-code": "^1.1.2",
     "fsm-event": "^2.1.0",
